# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Readbase",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/organizations": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Create Organization Endpoint",
        "operationId": "create_organization_endpoint_api_organizations_post",
        "parameters": [
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}": {
      "get": {
        "tags": [
          "organizations"
        ],
        "summary": "Get Organization Endpoint",
        "operationId": "get_organization_endpoint_api_organizations__org_id__get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/storage": {
      "patch": {
        "tags": [
          "organizations"
        ],
        "summary": "Update Organization Storage Endpoint",
        "operationId": "update_organization_storage_endpoint_api_organizations__org_id__storage_patch",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrganizationStorageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrganizationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/workspaces": {
      "post": {
        "tags": [
          "organizations"
        ],
        "summary": "Assign Workspace Organization Endpoint",
        "operationId": "assign_workspace_organization_endpoint_api_organizations__org_id__workspaces_post",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Org Id"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignWorkspaceOrganizationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceOrganizationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/me/integrations/teams/start": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Start Teams Connection",
        "operationId": "start_teams_connection_api_me_integrations_teams_start_get",
        "parameters": [
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/integrations/teams/callback": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Complete Teams Connection",
        "operationId": "complete_teams_connection_api_me_integrations_teams_callback_get",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "name": "state",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "name": "readbase_teams_oauth_state",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Teams Oauth State"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/me/integrations/teams": {
      "get": {
        "tags": [
          "teams"
        ],
        "summary": "Teams Connection",
        "operationId": "teams_connection_api_me_integrations_teams_get",
        "parameters": [
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamsConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "teams"
        ],
        "summary": "Disconnect Teams Connection",
        "operationId": "disconnect_teams_connection_api_me_integrations_teams_delete",
        "parameters": [
          {
            "name": "remove_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Remove Data"
            }
          },
          {
            "name": "readbase_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Readbase Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamsConnectionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssignWorkspaceOrganizationRequest": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          }
        },
        "type": "object",
        "required": [
          "workspace_id"
        ],
        "title": "AssignWorkspaceOrganizationRequest"
      },
...
      "CreateOrganizationRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "storage_root": {
            "type": "string",
            "title": "Storage Root"
          },
          "blob_backend": {
            "type": "string",
            "title": "Blob Backend",
            "default": "local"
          }
        },
        "type": "object",
        "required": [
          "name",
          "storage_root"
        ],
        "title": "CreateOrganizationRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrganizationResponse": {
        "properties": {
          "org_id": {
            "type": "string",
            "title": "Org Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "storage": {
            "$ref": "#/components/schemas/OrganizationStorageResponse"
          }
        },
        "type": "object",
        "required": [
          "org_id",
          "name",
          "role",
          "storage"
        ],
        "title": "OrganizationResponse"
      },
      "OrganizationStorageResponse": {
        "properties": {
          "blob_backend": {
            "type": "string",
            "title": "Blob Backend"
          },
          "storage_root": {
            "type": "string",
            "title": "Storage Root"
          }
        },
        "type": "object",
        "required": [
          "blob_backend",
          "storage_root"
        ],
        "title": "OrganizationStorageResponse"
      },
...
      "TeamsConnectionResponse": {
        "properties": {
          "connected": {
            "type": "boolean",
            "title": "Connected"
          },
          "configured": {
            "type": "boolean",
            "title": "Configured",
            "default": false
          },
          "microsoft_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Microsoft User Id"
          },
          "tenant_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tenant Id"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "user_principal_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Principal Name"
          },
          "mail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mail"
          },
          "scopes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scopes",
            "default": []
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/TeamsTeamResponse"
            },
            "type": "array",
            "title": "Teams",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "connected"
        ],
        "title": "TeamsConnectionResponse"
      },
      "TeamsTeamResponse": {
        "properties": {
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "web_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Web Url"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "display_name"
        ],
        "title": "TeamsTeamResponse"
      },
...
      "UpdateOrganizationStorageRequest": {
        "properties": {
          "storage_root": {
            "type": "string",
            "title": "Storage Root"
          },
          "blob_backend": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Blob Backend"
          }
        },
        "type": "object",
        "required": [
          "storage_root"
        ],
        "title": "UpdateOrganizationStorageRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WorkspaceOrganizationResponse": {
        "properties": {
          "workspace_id": {
            "type": "string",
            "title": "Workspace Id"
          },
          "organization_id": {
            "type": "string",
            "title": "Organization Id"
          }
        },
        "type": "object",
        "required": [
          "workspace_id",
          "organization_id"
        ],
        "title": "WorkspaceOrganizationResponse"
      },
...
    }
  }
}